requirements.txt
flask
pyodbc
anthropic
weasyprint
python-dotenv
